Add the python binary directory to `which`'s PATH
Appends the directory of the executing Python bin to the path of the
search for `which` function. This should help with a corner case
on clusters which don't export the PATH variable correctly on nodes
of the executing environment since the absolute Python binary's path
ensures proper imports.

Also cleaned up some unused imports and `__all__` statements.