Enable Azure CI with custom images for proprietary code
This PR creates an Azure Pipeline which uses a custom Docker image we (MolSSI)
maintain and have exclusive access to through Azure Container Registry.

On this image we can house non-public code which will enable us to test
Engine through integrated CI on these codes securely.

I will say, there are lots of shennanigans and chain of aobut 20-25 different
commits of trying different things which got squished into this, but I can
confirm it works. Once this is merged in, I will activate it on the
Azure side for Engine itself and we can see it in action.

Right now, the image only contains a static MOPAC binary, but we could expand it
to include packages, or make other images and set up a strategy of different
containers to run through the pipeline. Thanks to Jonathan Moussa (godotalgorithm)
for helping get the binary.