Get the scratch directory configuration before getting version
NWChem is an MPI parallel application, so it may not actually run
on the same node as that which is running QCEngine (this is the
case for Cray XC40s, which have MOM nodes from which you call
aprun to launch on compute nodes). Accordingly, we must apply some of
the job configuration information (e.g., scratch location) before
re-running.