Added Hessian support and improved gradient accuracy
NWChem only prints gradients to an accuracy of 6 decimal places to
stdout. The full-accuracy versions are stored in checkpoint file of
NWChem. To access the full-precision gradients, we access and print them
using a Python script that runs during the NWChem execution.

See discussion here:
http://www.nwchem-sw.org/index.php/Special:AWCforum/st/id2155

I also implemented Hessian support while I was at it