Allow NWChem to restart single-point computations
As written, the NWCHarness does not write the geometry to the input
file for restart calculations. This is great for Driver computations
because it prevents the computation from restarting from the initial
geometry. However, the geometry is required for single point calculations
(esp Hessian, Gradient) because we use the geometry to determine
how NWChem rotates the coordinate system so that we can adjust
any direction-dependant quantities back to the initial coordinate system.

This change is simple: I add a flag that is set by the "driver" computation
to tell the NWCHarness not to write the geometry for restarts.
For clarity, the NWCOptimizerHarness calls the NWChemHardness
to make input files